+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Nomenclature :
 
 - Added GET /me (controller, service and tests)
 - Added api-prefix
+- Updated register : checked if email used to login is existing
 
 #### Fixed