+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Nomenclature :
 - Updated register : checked if email used to login is existing
 - Added GET /properties : TDD (service and controller)
 - Added tables in liquibase : property , type and property_type (010), (011), (012)
+- Added POST /properties : TDD (service and controller)
+- Added GET /properties/:id and PUT /properties/:id (service and controller with TDD)
 
 #### Fixed